pyproject.toml
[project]
name = "concert-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.0",
    "fastapi>=0.115.0",
    "logfire>=2.8.1",
    "pydantic-ai>=1.57.0",
    "pydantic-ai-slim[duckduckgo,logfire]>=1.57.0",
    "sqladmin>=0.20.0",
    "sqlalchemy>=2.0.46",
    "html2text>=2024.2.26",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
concert-checker = "concert_checker.main:main"
concert-admin = "concert_checker.app.admin:serve"

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.15.2",
  "basedpyright>=1.38.1",
  "ipdb>=0.13.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = ["integration: tests that hit real external services (IMAP, APIs)"]
